require salt; capture and log errors from loader
